fix bugs
1. Fixed the remark "The number of coins is not updated in the Upgrade hero window"|| (the method was not called when opening the window)||
2. Fixed the remark "Mobs pass through the gate" ||Added a collider||
3. Fixed the remark "When restarting, the gate is immediately destroyed, at the same second"|| (the method was not called when opening the window), also added a View for visual control of the gate's HP||
4. Fixed the remark "And the cannon does not fire", ||added an additional event for launching the turret||
5. Fixed the remark "Fill imaga is not reset to delay ads"||, reset on exit was added, similar contact platforms were also checked||
6. Fixed the remark "The numbers were updated on the view, only after the update was taken, the coins were also worth 0." || (update methods were not called when opening the window)||
7. Fixed the remark "And it would also be good to clean up all mobs after restarting the level, otherwise it turns out that they just appeared, and they are already at the gate, they are hammering them, you don't even have time to react.", || added the logic of respawning enemies||
8. Fixed the remark "Moderation can be wrapped for the fact that the letters moved out in the word for Free. And it's not very obvious, I didn't notice at first." ||the layout has been corrected, the color has become red and the font is bold||
9. The remark "Reloading by 1 cartridge is tough" has been fixed|| the reloading of weapons has been accelerated||

+++ Assets/Scripts/SO/EnemyData.cs
@@ -18,9 +18,9 @@ namespace SO
         [Range(5, 15)] 
         public int AgroDistance = 5;
         [Range(1, 5)] 
-        public int AttackDistance = 1;
+        public float AttackDistance = 1;
         [Range(1, 15)] 
-        public int BossAttackDistance = 1;
+        public float BossAttackDistance = 1;
 
         [Header("Health simple enemies")] 
         [Range(1, 50)]
